gerri/robot/examples/zimmer/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zimmer"
version = "0.1.0"
description = "CLI tool for controlling Zimmer Gripper via Modbus TCP"
authors = [
  { name = "정윤철", email = "ycpiglet@gmail.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "click",
  "pymodbus",
  "python-dotenv"
]

# CLI entry point
[project.scripts]
zimmer = "zimmer.cli.main:cli"   # ✅ 패키지 이름 포함으로 변경

[tool.setuptools.packages.find]
where = ["src"]
include = ["zimmer*"]            # ✅ 패키지 이름으로 탐색